feat(app) refactoring create object
