+++ API/models/schemas/device_schema.json
@@ -84,7 +84,7 @@
           "template": "huawei-xxxxxx",
           "type": "blade",
           "vendor": "Huawei",
-          "weightKg": 1.81
+          "weightKg": "1.81"
       },
       "category": "device",
       "description": [
