+++ APP/lib/widgets/select_objects/object_popup.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:ogree_app/common/api_backend.dart';
@@ -7,29 +8,37 @@ import 'package:ogree_app/common/definitions.dart';
 import 'package:ogree_app/common/snackbar.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:ogree_app/common/theme.dart';
+import 'package:ogree_app/models/tag.dart';
 import 'package:ogree_app/widgets/actionbtn_row.dart';
 
-class CreateObjectPopup extends StatefulWidget {
+class ObjectPopup extends StatefulWidget {
   Function() parentCallback;
-  String? parentId;
   String? objId;
   Namespace namespace;
-  CreateObjectPopup(
+  String? parentId;
+  ObjectPopup(
       {super.key,
       required this.parentCallback,
       required this.namespace,
-      this.parentId,
-      this.objId});
+      this.objId,
+      this.parentId});
 
   @override
-  State<CreateObjectPopup> createState() => _CreateObjectPopupState();
+  State<ObjectPopup> createState() => _ObjectPopupState();
 }
 
 enum PhyCategories { site, building, room, rack, device, group }
 
 enum OrgCategories { domain }
 
-enum LogCategories { group, layer, obj_template, room_template, tag }
+enum LogCategories {
+  group,
+  layer,
+  obj_template,
+  room_template,
+  bldg_template,
+  tag
+}
 
 Map<Namespace, List<String>> objsByNamespace = {
   Namespace.Physical: PhyCategories.values.map((e) => e.name).toList(),
@@ -37,19 +46,32 @@ Map<Namespace, List<String>> objsByNamespace = {
   Namespace.Logical: LogCategories.values.map((e) => e.name).toList(),
 };
 
-class _CreateObjectPopupState extends State<CreateObjectPopup> {
+class _ObjectPopupState extends State<ObjectPopup> {
   final _formKey = GlobalKey<FormState>();
   bool _isSmallDisplay = false;
-  String _objCategory = PhyCategories.site.name;
+  String _objCategory = LogCategories.group.name;
   String _objId = "";
   List<Widget> attributesRows = [];
   List<String> attributes = [];
   Map<String, List<String>> categoryAttrs = {};
+  Map<String, Map<String, String>> examplesAttrs = {};
   List<String> domainList = [];
   Map<String, dynamic> createObjData = {};
   Map<String, dynamic> editObjData = {};
   Map<String, String> createObjDataAttrs = {};
   bool _isEdit = false;
+
+  // Tags
+  String? _tagSlug;
+  String? _tagDescription;
+  String? _tagColor;
+  Color? _localColor;
+  Tag? tag;
+  PlatformFile? _loadedImage;
+
+  // Templates
+  PlatformFile? _loadedFile;
+  String? _loadFileResult;
 
   @override
   void initState() {
@@ -57,6 +79,8 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
     if (widget.parentId != null) {
       createObjData["parentId"] = widget.parentId;
       if (widget.namespace == Namespace.Organisational) {
+        _objCategory = OrgCategories.domain.name;
+      } else if (widget.namespace == Namespace.Logical) {
         _objCategory = OrgCategories.domain.name;
       } else {
         switch (".".allMatches(widget.parentId!).length) {
@@ -82,11 +106,11 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
           _objCategory = PhyCategories.site.name;
       }
     }
+
     if (widget.objId != null && widget.objId!.isNotEmpty) {
       _isEdit = true;
       _objId = widget.objId!;
     }
-    print("IS EDIT " + _isEdit.toString());
   }
 
   @override
@@ -100,13 +124,14 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
           if (categoryAttrs.isEmpty) {
             return const Center(child: CircularProgressIndicator());
           }
-          attributes = categoryAttrs[_objCategory]!;
+          if (widget.namespace == Namespace.Physical) {
+            attributes = categoryAttrs[_objCategory]!;
+          }
           return Center(
             child: Container(
               width: 500,
-              constraints: BoxConstraints(
-                  maxHeight:
-                      _objCategory != OrgCategories.domain.name ? 590 : 470),
+              constraints:
+                  BoxConstraints(maxHeight: getPopupHeightByCategory()),
               margin: const EdgeInsets.symmetric(horizontal: 20),
               decoration: PopupDecoration,
               child: Padding(
@@ -117,190 +142,102 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
                         child: Builder(
                       builder: (context) => Scaffold(
                         backgroundColor: Colors.white,
-                        body: ListView(
-                          padding: EdgeInsets.zero,
-                          children: [
-                            Center(
-                              child: Text(
-                                _isEdit
-                                    ? "Modifier l'objet"
-                                    : "Créer un nouveau objet",
-                                style:
-                                    Theme.of(context).textTheme.headlineMedium,
-                              ),
-                            ),
-                            // const Divider(height: 45),
-                            const SizedBox(height: 20),
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                Text("Type d'objet :"),
-                                const SizedBox(width: 20),
-                                SizedBox(
-                                  height: 35,
-                                  width: 147,
-                                  child: DropdownButtonFormField<String>(
-                                    borderRadius: BorderRadius.circular(12.0),
-                                    decoration: GetFormInputDecoration(
-                                      false,
-                                      null,
-                                      icon: Icons.bookmark,
-                                    ),
-                                    value: _objCategory,
-                                    items: objsByNamespace[widget.namespace]!
-                                        .map<DropdownMenuItem<String>>(
-                                            (String value) {
-                                      return DropdownMenuItem<String>(
-                                        value: value,
-                                        child: Text(
-                                          value,
-                                          overflow: TextOverflow.ellipsis,
-                                        ),
-                                      );
-                                    }).toList(),
-                                    onChanged: _isEdit
-                                        ? null
-                                        : (String? value) {
-                                            setState(() {
-                                              _objCategory = value!;
-                                            });
-                                          },
-                                  ),
+                        body: SingleChildScrollView(
+                          child: Column(
+                            // padding: EdgeInsets.zero,
+                            children: [
+                              Center(
+                                child: Text(
+                                  _isEdit
+                                      ? (_objCategory.contains("template")
+                                          ? "Visualiser template"
+                                          : "Modifier l'objet")
+                                      : "Créer un nouveau objet",
+                                  style: Theme.of(context)
+                                      .textTheme
+                                      .headlineMedium,
                                 ),
-                              ],
-                            ),
-                            const SizedBox(height: 10),
-                            _objCategory != PhyCategories.site.name
-                                ? getFormField(
-                                    save: (newValue) {
-                                      if (newValue != null &&
-                                          newValue.isNotEmpty) {
-                                        createObjData["parentId"] = newValue;
-                                      }
-                                    },
-                                    label: "Parent ID",
-                                    icon: Icons.family_restroom,
-                                    initial: createObjData["parentId"])
-                                : Container(),
-                            getFormField(
-                                save: (newValue) =>
-                                    createObjData["name"] = newValue,
-                                label: "Name",
-                                icon: Icons.edit,
-                                initial: createObjData["name"]),
-                            _objCategory != OrgCategories.domain.name
-                                ? (domainList.isEmpty
-                                    ? getFormField(
-                                        save: (newValue) =>
-                                            createObjData["domain"] = newValue,
-                                        label: "Domain",
-                                        icon: Icons.edit,
-                                        initial: createObjData["domain"])
-                                    : domainAutoFillField())
-                                : Container(),
-                            getFormField(
-                                save: (newValue) =>
-                                    createObjData["description"] = [newValue],
-                                label: "Description",
-                                icon: Icons.edit,
-                                shouldValidate: false,
-                                initial: createObjData["description"] != null &&
-                                        List<String>.from(
-                                                createObjData["description"])
-                                            .isNotEmpty
-                                    ? createObjData["description"][0]
-                                    : null),
-                            _objCategory != OrgCategories.domain.name
-                                ? getFormField(
-                                    save: (newValue) {
-                                      var tags = newValue!
-                                          .replaceAll(" ", "")
-                                          .split(",");
-                                      if (!(tags.length == 1 &&
-                                          tags.first == "")) {
-                                        createObjData["tags"] = tags;
-                                      }
-                                    },
-                                    label: "Tags",
-                                    icon: Icons.tag_sharp,
-                                    shouldValidate: false,
-                                    initial: createObjData["tags"]
-                                        ?.toString()
-                                        .substring(
-                                            1,
-                                            createObjData["tags"]
-                                                    .toString()
-                                                    .length -
-                                                1))
-                                : Container(),
-                            Padding(
-                              padding: const EdgeInsets.only(
-                                  top: 4.0, left: 6, bottom: 6),
-                              child: Text("Attributes:"),
-                            ),
-                            SizedBox(
-                              height: (attributes.length ~/ 2 +
-                                      attributes.length % 2) *
-                                  60,
-                              child: GridView.count(
-                                physics: NeverScrollableScrollPhysics(),
-                                childAspectRatio: 3.5,
-                                shrinkWrap: true,
-                                padding: EdgeInsets.only(left: 4),
-                                // Create a grid with 2 columns
-                                crossAxisCount: 2,
-                                children:
-                                    List.generate(attributes.length, (index) {
-                                  print(createObjDataAttrs[
-                                      attributes[index].replaceFirst("*", "")]);
-                                  return getFormField(
-                                      save: (newValue) {
-                                        if (newValue != null &&
-                                            newValue.isNotEmpty) {
-                                          createObjDataAttrs[attributes[index]
-                                                  .replaceFirst("*", "")] =
-                                              newValue;
-                                        }
-                                      },
-                                      label: attributes[index],
-                                      icon: Icons.tag_sharp,
-                                      isCompact: true,
-                                      shouldValidate:
-                                          attributes[index].contains("*"),
-                                      initial: createObjDataAttrs[
-                                          attributes[index]
-                                              .replaceFirst("*", "")]);
-                                }),
                               ),
-                            ),
-                            Padding(
-                              padding: const EdgeInsets.only(left: 4),
-                              child: Column(children: attributesRows),
-                            ),
-                            Padding(
-                              padding: const EdgeInsets.only(left: 6),
-                              child: Align(
-                                alignment: Alignment.bottomLeft,
-                                child: TextButton.icon(
-                                    onPressed: () => setState(() {
-                                          attributesRows.add(addCustomAttrRow(
-                                              attributesRows.length));
-                                        }),
-                                    icon: const Icon(Icons.add),
-                                    label: Text("Attribute")),
+                              // const Divider(height: 45),
+                              const SizedBox(height: 20),
+                              Row(
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Text("Type d'objet :"),
+                                  const SizedBox(width: 20),
+                                  SizedBox(
+                                    height: 35,
+                                    width: 147,
+                                    child: DropdownButtonFormField<String>(
+                                      isExpanded: true,
+                                      borderRadius: BorderRadius.circular(12.0),
+                                      decoration: GetFormInputDecoration(
+                                        false,
+                                        null,
+                                        icon: Icons.bookmark,
+                                      ),
+                                      value: _objCategory,
+                                      items: objsByNamespace[widget.namespace]!
+                                          .map<DropdownMenuItem<String>>(
+                                              (String value) {
+                                        return DropdownMenuItem<String>(
+                                          value: value,
+                                          child: Text(
+                                            value,
+                                            overflow: TextOverflow.ellipsis,
+                                          ),
+                                        );
+                                      }).toList(),
+                                      onChanged: _isEdit
+                                          ? null
+                                          : (String? value) {
+                                              setState(() {
+                                                _objCategory = value!;
+                                              });
+                                            },
+                                    ),
+                                  ),
+                                ],
                               ),
-                            ),
-                            const SizedBox(height: 12),
-                            ActionBtnRow(
-                              isEdit: _isEdit,
-                              submitCreate: () =>
-                                  submitCreateObject(localeMsg, context),
-                              submitModify: () =>
-                                  submitModifyObject(localeMsg, context),
-                              submitDelete: () =>
-                                  submitDeleteObject(localeMsg, context),
-                            )
-                          ],
+                              const SizedBox(height: 10),
+                              SizedBox(
+                                  height: getFormHeightByCategory(),
+                                  child: getFormByCategory(
+                                      _objCategory, localeMsg)),
+
+                              const SizedBox(height: 12),
+                              ActionBtnRow(
+                                isEdit: _isEdit,
+                                onlyDelete: _isEdit &&
+                                    _objCategory.contains("template"),
+                                submitCreate: () {
+                                  if (_objCategory == LogCategories.tag.name) {
+                                    return onActionBtnPressed(localeMsg);
+                                  } else if (_objCategory
+                                      .contains("template")) {
+                                    return submitCreateTemplate(
+                                        localeMsg, context);
+                                  } else {
+                                    return submitCreateObject(
+                                        localeMsg, context);
+                                  }
+                                },
+                                submitModify: () {
+                                  if (_objCategory == LogCategories.tag.name) {
+                                    return onActionBtnPressed(localeMsg);
+                                  } else if (_objCategory
+                                      .contains("template")) {
+                                    return;
+                                  } else {
+                                    return submitModifyObject(
+                                        localeMsg, context);
+                                  }
+                                },
+                                submitDelete: () =>
+                                    submitDeleteObject(localeMsg, context),
+                              )
+                            ],
+                          ),
                         ),
                       ),
                     ))),
@@ -308,6 +245,30 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
             ),
           );
         });
+  }
+
+  double getPopupHeightByCategory() {
+    if (widget.namespace == Namespace.Physical ||
+        _objCategory == LogCategories.group.name) {
+      return 585;
+    } else if (widget.namespace == Namespace.Organisational) {
+      return 470;
+    } else {
+      // Logical, except group
+      return 390;
+    }
+  }
+
+  double getFormHeightByCategory() {
+    if (widget.namespace == Namespace.Physical ||
+        _objCategory == LogCategories.group.name) {
+      return 415;
+    } else if (widget.namespace == Namespace.Organisational) {
+      return 300;
+    } else {
+      // Logical, except group
+      return 220;
+    }
   }
 
   getExternalAssets() async {
@@ -319,7 +280,13 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
   }
 
   readJsonAssets() async {
-    List<String> objects = objsByNamespace[widget.namespace]!;
+    List<String> objects = [LogCategories.group.name];
+    if (widget.namespace == Namespace.Physical) {
+      objects = objsByNamespace[widget.namespace]!;
+    } else if (widget.namespace == Namespace.Organisational) {
+      objects = objsByNamespace[widget.namespace]!;
+    }
+
     for (var obj in objects) {
       print(obj);
       // read JSON schema
@@ -344,8 +311,13 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
           }
         }
         categoryAttrs[obj]!.sort((a, b) => a.compareTo(b));
+        var examples = List<Map<String, dynamic>>.from(jsonResult["examples"]);
+        examplesAttrs[obj] =
+            Map<String, String>.from(examples[0]["attributes"]);
+        print(examplesAttrs);
       }
     }
+    print(examplesAttrs);
   }
 
   getDomains() async {
@@ -367,18 +339,66 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
 
   getObject() async {
     final messenger = ScaffoldMessenger.of(context);
-    var result = await fetchObject(_objId);
-    switch (result) {
-      case Success(value: final value):
-        createObjData = value;
-        createObjDataAttrs =
-            Map<String, String>.from(createObjData["attributes"]);
-        print(createObjDataAttrs);
-      case Failure(exception: final exception):
-        showSnackBar(messenger, exception.toString(), isError: true);
-        if (context.mounted) Navigator.pop(context);
-        return;
+    var errMsg = "";
+
+    for (var keyId in ["id", "slug"]) {
+      var result = await fetchObject(_objId, idKey: keyId);
+      switch (result) {
+        case Success(value: final value):
+          if (widget.namespace == Namespace.Logical) {
+            if (["room", "building", "device", "rack", "generic"]
+                .contains(value["category"])) {
+              switch (value["category"]) {
+                case "room":
+                  _objCategory = LogCategories.room_template.name;
+                  break;
+                case "building":
+                  _objCategory = LogCategories.bldg_template.name;
+                  break;
+                default:
+                  _objCategory = LogCategories.obj_template.name;
+              }
+              var encoder = new JsonEncoder.withIndent("     ");
+              _loadFileResult = encoder.convert(value);
+              // _loadFileResult = value.toString();
+            } else {
+              if (value["applicability"] != null) {
+                _objCategory = LogCategories.layer.name;
+                createObjData = value;
+                createObjDataAttrs =
+                    Map<String, String>.from(createObjData["filters"]);
+                for (var attr in createObjDataAttrs.entries) {
+                  attributesRows.add(addCustomAttrRow(attributesRows.length,
+                      givenAttrName: attr.key, givenAttrValue: attr.value));
+                }
+              } else if (value["category"] == null) {
+                _objCategory = LogCategories.tag.name;
+                tag = Tag.fromMap(value);
+                _localColor = Color(int.parse("0xFF${tag!.color}"));
+              } else {
+                createObjData = value;
+                createObjDataAttrs =
+                    Map<String, String>.from(createObjData["attributes"]);
+              }
+
+              print("GOT OBJECT");
+              print(createObjDataAttrs);
+            }
+          } else {
+            createObjData = value;
+            createObjDataAttrs =
+                Map<String, String>.from(createObjData["attributes"]);
+            print("HEEEERE");
+            print(createObjDataAttrs);
+          }
+
+          return;
+        case Failure(exception: final exception):
+          errMsg = exception.toString();
+      }
     }
+    showSnackBar(messenger, errMsg, isError: true);
+    if (context.mounted) Navigator.pop(context);
   }
 
   domainAutoFillField() {
@@ -444,9 +464,15 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
     );
   }
 
-  addCustomAttrRow(int rowIdx, {bool useDefaultValue = true}) {
+  addCustomAttrRow(int rowIdx,
+      {bool useDefaultValue = true,
+      String? givenAttrName,
+      String? givenAttrValue}) {
     return StatefulBuilder(builder: (context, localSetState) {
       String? attrName;
+      if (givenAttrName != null) {
+        attrName = givenAttrName;
+      }
       return Padding(
         padding: const EdgeInsets.only(top: 2.0),
         child: SizedBox(
@@ -460,7 +486,8 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
                     save: (newValue) => attrName = newValue,
                     label: "Attribute",
                     icon: Icons.tag_sharp,
-                    isCompact: true),
+                    isCompact: true,
+                    initial: givenAttrName),
               ),
               Padding(
                 padding: EdgeInsets.only(right: 6),
@@ -476,14 +503,17 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
                         createObjDataAttrs[attrName!] = newValue!,
                     label: "Value",
                     icon: Icons.tag_sharp,
-                    isCompact: true),
+                    isCompact: true,
+                    initial: givenAttrValue),
               ),
               IconButton(
                   padding: const EdgeInsets.only(bottom: 6),
                   constraints: const BoxConstraints(),
                   iconSize: 14,
-                  onPressed: () =>
-                      setState(() => attributesRows.removeAt(rowIdx)),
+                  onPressed: () {
+                    setState(() => attributesRows.removeAt(rowIdx));
+                    createObjDataAttrs.remove(attrName);
+                  },
                   icon: Icon(
                     Icons.delete,
                     color: Colors.red.shade400,
@@ -495,13 +525,357 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
     });
   }
 
+  getFormByCategory(String category, AppLocalizations localeMsg) {
+    print("get form by " + category);
+    if (widget.namespace == Namespace.Physical ||
+        widget.namespace == Namespace.Organisational ||
+        category == LogCategories.group.name) {
+      return getObjectForm();
+    } else if (category == LogCategories.layer.name) {
+      return getLayerForm();
+    } else if (category == LogCategories.tag.name) {
+      return getTagForm();
+    } else {
+      //templates
+      return getTemplatesForm(localeMsg);
+    }
+  }
+
+  getObjectForm() {
+    attributes = categoryAttrs[_objCategory]!;
+    print(attributes);
+    return ListView(
+      padding: EdgeInsets.zero,
+      children: [
+        _objCategory != PhyCategories.site.name
+            ? getFormField(
+                save: (newValue) {
+                  if (newValue != null && newValue.isNotEmpty) {
+                    createObjData["parentId"] = newValue;
+                  }
+                },
+                label: "Parent ID",
+                icon: Icons.family_restroom,
+                initial: createObjData["parentId"])
+            : Container(),
+        getFormField(
+            save: (newValue) => createObjData["name"] = newValue,
+            label: "Name",
+            icon: Icons.edit,
+            initial: createObjData["name"]),
+        _objCategory != OrgCategories.domain.name
+            ? (domainList.isEmpty
+                ? getFormField(
+                    save: (newValue) => createObjData["domain"] = newValue,
+                    label: "Domain",
+                    icon: Icons.edit,
+                    initial: createObjData["domain"])
+                : domainAutoFillField())
+            : Container(),
+        getFormField(
+            save: (newValue) => createObjData["description"] = [newValue],
+            label: "Description",
+            icon: Icons.edit,
+            shouldValidate: false,
+            initial: createObjData["description"] != null &&
+                    List<String>.from(createObjData["description"]).isNotEmpty
+                ? createObjData["description"][0]
+                : null),
+        _objCategory != OrgCategories.domain.name
+            ? getFormField(
+                save: (newValue) {
+                  var tags = newValue!.replaceAll(" ", "").split(",");
+                  if (!(tags.length == 1 && tags.first == "")) {
+                    createObjData["tags"] = tags;
+                  }
+                },
+                label: "Tags",
+                icon: Icons.tag_sharp,
+                shouldValidate: false,
+                initial: createObjData["tags"]
+                    ?.toString()
+                    .substring(1, createObjData["tags"].toString().length - 1))
+            : Container(),
+        Padding(
+          padding: const EdgeInsets.only(top: 4.0, left: 6, bottom: 6),
+          child: Text("Attributes:"),
+        ),
+        SizedBox(
+          height: (attributes.length ~/ 2 + attributes.length % 2) * 60,
+          child: GridView.count(
+            physics: NeverScrollableScrollPhysics(),
+            childAspectRatio: 3.5,
+            shrinkWrap: true,
+            padding: EdgeInsets.only(left: 4),
+            // Create a grid with 2 columns
+            crossAxisCount: 2,
+            children: List.generate(attributes.length, (index) {
+              return getFormField(
+                  tipStr: examplesAttrs[_objCategory]
+                          ?[attributes[index].replaceFirst("*", "")] ??
+                      "",
+                  save: (newValue) {
+                    if (newValue != null && newValue.isNotEmpty) {
+                      createObjDataAttrs[
+                          attributes[index].replaceFirst("*", "")] = newValue;
+                    }
+                  },
+                  label: attributes[index],
+                  icon: Icons.tag_sharp,
+                  isCompact: true,
+                  shouldValidate: attributes[index].contains("*"),
+                  initial: createObjDataAttrs[
+                      attributes[index].replaceFirst("*", "")]);
+            }),
+          ),
+        ),
+        Padding(
+          padding: const EdgeInsets.only(left: 4),
+          child: Column(children: attributesRows),
+        ),
+        Padding(
+          padding: const EdgeInsets.only(left: 6),
+          child: Align(
+            alignment: Alignment.bottomLeft,
+            child: TextButton.icon(
+                onPressed: () => setState(() {
+                      attributesRows
+                          .add(addCustomAttrRow(attributesRows.length));
+                    }),
+                icon: const Icon(Icons.add),
+                label: Text("Attribute")),
+          ),
+        ),
+      ],
+    );
+  }
+
+  getLayerForm() {
+    return ListView(padding: EdgeInsets.zero, children: [
+      getFormField(
+          save: (newValue) => createObjData["slug"] = newValue,
+          label: "Name",
+          icon: Icons.edit,
+          initial: createObjData["slug"]),
+      getFormField(
+          save: (newValue) => createObjData["applicability"] = newValue,
+          label: "Applicability",
+          icon: Icons.edit,
+          initial: createObjData["applicability"]),
+      Padding(
+          padding: const EdgeInsets.only(top: 4.0, left: 6, bottom: 6),
+          child: Text("Filters:")),
+      Padding(
+        padding: const EdgeInsets.only(left: 4),
+        child: Column(children: attributesRows),
+      ),
+      Padding(
+        padding: const EdgeInsets.only(left: 6),
+        child: Align(
+          alignment: Alignment.bottomLeft,
+          child: TextButton.icon(
+              onPressed: () => setState(() {
+                    attributesRows.add(addCustomAttrRow(attributesRows.length));
+                  }),
+              icon: const Icon(Icons.add),
+              label: Text("Filter")),
+        ),
+      ),
+    ]);
+  }
+
+  getTagForm() {
+    final localeMsg = AppLocalizations.of(context)!;
+    return ListView(
+      padding: EdgeInsets.zero,
+      children: [
+        getFormField(
+            save: (newValue) => _tagSlug = newValue,
+            label: "Slug",
+            icon: Icons.auto_awesome_mosaic,
+            initial: _isEdit ? tag!.slug : null),
+        getFormField(
+            save: (newValue) => _tagDescription = newValue,
+            label: "Description",
+            icon: Icons.auto_awesome_mosaic,
+            initial: _isEdit ? tag!.description : null),
+        getFormField(
+            save: (newValue) => _tagColor = newValue,
+            label: localeMsg.color,
+            icon: Icons.circle,
+            formatters: [
+              FilteringTextInputFormatter.allow(RegExp(r'[0-9a-fA-F]'))
+            ],
+            isColor: true,
+            initial: _isEdit ? tag!.color : null),
+        Padding(
+          padding: const EdgeInsets.only(top: 8.0, bottom: 8),
+          child: Wrap(
+            alignment: WrapAlignment.end,
+            crossAxisAlignment: WrapCrossAlignment.center,
+            children: [
+              _loadedImage != null || (_isEdit && tag!.image != "")
+                  ? IconButton(
+                      padding: const EdgeInsets.all(4),
+                      constraints: const BoxConstraints(),
+                      iconSize: 14,
+                      onPressed: () {
+                        setState(() {
+                          _loadedImage = null;
+                          tag!.image = "";
+                        });
+                      },
+                      icon: const Icon(
+                        Icons.cancel_outlined,
+                      ))
+                  : Container(),
+              Padding(
+                padding: const EdgeInsets.only(right: 20),
+                child: _loadedImage == null
+                    ? (_isEdit && tag!.image != ""
+                        ? Image.network(
+                            tenantUrl + tag!.image,
+                            height: 40,
+                          )
+                        : Container())
+                    : Image.memory(
+                        _loadedImage!.bytes!,
+                        height: 40,
+                      ),
+              ),
+              ElevatedButton.icon(
+                  onPressed: () async {
+                    FilePickerResult? result = await FilePicker.platform
+                        .pickFiles(
+                            type: FileType.custom,
+                            allowedExtensions: ["png", "jpg", "jpeg", "webp"],
+                            withData: true);
+                    if (result != null) {
+                      setState(() {
+                        _loadedImage = result.files.single;
+                      });
+                    }
+                  },
+                  icon: const Icon(Icons.download),
+                  label: Text(
+                      _isSmallDisplay ? "Image" : "${localeMsg.select} image")),
+            ],
+          ),
+        )
+      ],
+    );
+  }
+
+  getTemplatesForm(AppLocalizations localeMsg) {
+    return Center(
+      child: ListView(shrinkWrap: true, children: [
+        _loadFileResult == null
+            ? Align(
+                child: ElevatedButton.icon(
+                    onPressed: () async {
+                      FilePickerResult? result = await FilePicker.platform
+                          .pickFiles(
+                              type: FileType.custom,
+                              allowedExtensions: ["json"],
+                              withData: true);
+                      if (result != null) {
+                        setState(() {
+                          _loadedFile = result.files.single;
+                        });
+                      }
+                    },
+                    icon: const Icon(Icons.download),
+                    label: Text(localeMsg.selectJSON)),
+              )
+            : Container(),
+        _loadedFile != null
+            ? Padding(
+                padding: const EdgeInsets.only(top: 8.0, bottom: 8.0),
+                child: Align(
+                  child: Text(localeMsg.fileLoaded(_loadedFile!.name)),
+                ),
+              )
+            : Container(),
+        _loadFileResult != null
+            ? Container(
+                color: Colors.black,
+                child: Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: Text(
+                    _loadFileResult!,
+                    style: const TextStyle(color: Colors.white),
+                  ),
+                ),
+              )
+            : Container(),
+      ]),
+    );
+  }
+
+  onActionBtnPressed(AppLocalizations localeMsg) async {
+    final messenger = ScaffoldMessenger.of(context);
+    if (_formKey.currentState!.validate()) {
+      _formKey.currentState!.save();
+      var newTag = Tag(
+          slug: _tagSlug!,
+          description: _tagDescription!,
+          color: _tagColor!,
+          image: _loadedImage != null
+              ? "data:image/png;base64,${base64Encode(_loadedImage!.bytes!)}"
+              : "");
+      Result result;
+      if (_isEdit) {
+        var newTagMap = newTag.toMap();
+        if (_loadedImage == null && tag!.image != "") {
+          newTagMap.remove("image"); // patch and keep old one
+        }
+        result = await updateTag(tag!.slug, newTagMap);
+      } else {
+        result = await createTag(newTag);
+      }
+      switch (result) {
+        case Success():
+          widget.parentCallback();
+          showSnackBar(messenger,
+              "${_isEdit ? localeMsg.modifyOK : localeMsg.createOK} 🥳",
+              isSuccess: true);
+          if (context.mounted) Navigator.of(context).pop();
+        case Failure(exception: final exception):
+          showSnackBar(messenger, exception.toString(), isError: true);
+      }
+    }
+  }
+
+  submitCreateTemplate(
+      AppLocalizations localeMsg, BuildContext popupContext) async {
+    final messenger = ScaffoldMessenger.of(context);
+    final errorMessenger = ScaffoldMessenger.of(popupContext);
+    if (_loadedFile == null) {
+      showSnackBar(messenger, localeMsg.mustSelectJSON);
+    } else {
+      var result = await createTemplate(_loadedFile!.bytes!, _objCategory);
+      switch (result) {
+        case Success():
+          widget.parentCallback();
+          showSnackBar(messenger, localeMsg.createOK, isSuccess: true);
+          if (context.mounted) Navigator.of(context).pop();
+        case Failure(exception: final exception):
+          showSnackBar(errorMessenger, exception.toString(), isError: true);
+      }
+    }
+  }
+
   submitCreateObject(
       AppLocalizations localeMsg, BuildContext popupContext) async {
     if (_formKey.currentState!.validate()) {
       _formKey.currentState!.save();
 
-      createObjData["category"] = _objCategory;
-      createObjData["attributes"] = createObjDataAttrs;
+      if (_objCategory == LogCategories.layer.name) {
+        createObjData["filters"] = createObjDataAttrs;
+      } else {
+        createObjData["category"] = _objCategory;
+        createObjData["attributes"] = createObjDataAttrs;
+      }
       print(createObjData);
 
       final messenger = ScaffoldMessenger.of(context);
@@ -523,8 +897,13 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
     if (_formKey.currentState!.validate()) {
       _formKey.currentState!.save();
 
-      createObjData["category"] = _objCategory;
-      createObjData["attributes"] = createObjDataAttrs;
+      if (_objCategory == LogCategories.layer.name) {
+        createObjData["filters"] = createObjDataAttrs;
+      } else {
+        createObjData["category"] = _objCategory;
+        createObjData["attributes"] = createObjDataAttrs;
+      }
+
       createObjData.remove("lastUpdated");
       createObjData.remove("createdDate");
       createObjData.remove("id");
@@ -537,6 +916,7 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
         case Success():
           widget.parentCallback();
           showSnackBar(messenger, localeMsg.modifyOK, isSuccess: true);
+          if (context.mounted) Navigator.of(context).pop();
         case Failure(exception: final exception):
           showSnackBar(messenger, exception.toString(), isError: true);
       }
@@ -569,25 +949,46 @@ class _CreateObjectPopupState extends State<CreateObjectPopup> {
       List<TextInputFormatter>? formatters,
       String? initial,
       bool shouldValidate = true,
-      bool isCompact = false}) {
+      bool isColor = false,
+      bool isCompact = false,
+      String tipStr = ""}) {
     return Padding(
       padding: FormInputPadding,
-      child: TextFormField(
-        initialValue: initial,
-        onSaved: (newValue) => save(newValue),
-        validator: (text) {
-          if (shouldValidate) {
-            if (text == null || text.isEmpty) {
-              return AppLocalizations.of(context)!.mandatoryField;
+      child: Tooltip(
+        message: tipStr != "" ? "Example: $tipStr" : "",
+        child: TextFormField(
+          initialValue: initial,
+          onChanged: isColor
+              ? (value) {
+                  if (value.length == 6) {
+                    setState(() {
+                      _localColor = Color(int.parse("0xFF$value"));
+                    });
+                  } else {
+                    setState(() {
+                      _localColor = null;
+                    });
+                  }
+                }
+              : null,
+          onSaved: (newValue) => save(newValue),
+          validator: (text) {
+            if (shouldValidate) {
+              if (text == null || text.isEmpty) {
+                return AppLocalizations.of(context)!.mandatoryField;
+              }
             }
-          }
-          return null;
-        },
-        inputFormatters: formatters,
-        decoration: GetFormInputDecoration(_isSmallDisplay | isCompact, label,
-            prefixText: prefix, suffixText: suffix, icon: icon),
-        cursorWidth: 1.3,
-        style: const TextStyle(fontSize: 14),
+            return null;
+          },
+          inputFormatters: formatters,
+          decoration: GetFormInputDecoration(_isSmallDisplay | isCompact, label,
+              prefixText: prefix,
+              suffixText: suffix,
+              icon: icon,
+              iconColor: isColor ? _localColor : null),
+          cursorWidth: 1.3,
+          style: const TextStyle(fontSize: 14),
+        ),
       ),
     );
   }
